Refactor: Enhance route scalability and update URL provision logic
- Update the routing path to /api/v1/dbx/services from /dbx to improve scalability and clarity in route naming, aligning with RESTful API design principles.
- Modify the providedUrlController to getFixedUrl, refining the function to offer more comprehensive error handling and to return JSON responses instead of redirecting.
- Implement initial resource ID validation in the getFixedUrl function, returning a 400 status code with an error message if the resource ID is missing from the request.
- Enhance resource existence check within getFixedUrl, now returning a 404 status code with an appropriate error message if the specified resource cannot be found.
- Adjust error handling in the getFixedUrl function to provide a 500 status code response with a generic server error message in case of unexpected failures, improving the API's robustness and reliability.

This refactor addresses the need for more scalable route management and optimizes the logic for providing graphic resource URLs, enhancing the API's maintainability and usability.
